refactor actions to not use the KShortcutsDialog from KXmlGui
Actions are now created with an Instantiator from actionsModel, when an action is triggered
the actionsModel signalEmitter function is called with the action's objectName.
The objectName is used to invoke the signal corresponding to the action, these signals are handled in
a Connections item in the actions.qml file.
Actions are also added to the actions QQmlPropertyMap so they are easily accessible.

The actionsModel is also used for changing the actions shortcuts in ShortcutsSettings.qml