Fix blur effect transition
Blur effect does not follow the playlist's visible/hidden transition. Instead, it always renders the full rectangle. Making it look like not a transparent window, but a viewport. Which is not intended I think? Also, it does not consider whether the toolbar is visible or not, making the render squished a little.
- Blur rectangle now follows the transition
- Blur rectangle now respects the toolbar
- Added commented code in case of a future footbar addition