Fix edge border position when playlist is on left side
This includes two commits. 

First one fixes a visual bug for left aligned playlist. `playlistEdgeBorder` positions itself out of the ScrollView. This makes it visually have 2 pixels on the right side when it should have 1 pixel. A little inconsistent and makes the big drag handler looks visually not-centered (this is how I noticed it).

Note that there is another solution, if you prefer:

```diff
diff --git a/src/qml/PlayList.qml b/src/qml/PlayList.qml
index 09940e8c..5924419e 100644
--- a/src/qml/PlayList.qml
+++ b/src/qml/PlayList.qml
@@ -436,7 +436,7 @@ Page {
         Rectangle {
             id: playlistEdgeBorder
 
-            x: PlaylistSettings.position === "right" ? 0 : parent.width
+            x: PlaylistSettings.position === "right" ? -width : parent.width
             y: -root.implicitHeaderHeight
             z: 30
             width: 1
@@ -488,11 +488,6 @@ Page {
 
             z: 20
             anchors.fill: parent
-            anchors {
-                leftMargin: playlistEdgeBorder.width
-                rightMargin: playlistEdgeBorder.width
-            }
-
             ScrollBar.horizontal.policy: ScrollBar.AlwaysOff
 
             ListView {
```

This also removes the margins. I put them in the first place because item delegates were obstructing the edge borders. The patch above puts the border outside of the playlist, so it is never covered. 

Second commit makes the big drag handler a rounded rect and adds borders to it (for varying colors).

PS: If you don't like the second one, feel free to close this MR and push any of the fixes yourself. I don't mind.