Raise CMake minimum to 3.16...3.30 range
ECM Find* and Qt6 CMake modules already require 3.16, so it is a good
base line. CMake 3.16 was released in 2019. At the same time, defining
the range to 3.30 means the build system is already being tested with
all of the Policy changes up to build time CMake version <=3.30 enacted.
3.30 is picked as CMP0175 (in 3.31) introduced an error in <ECM-6.20
KDEClangFormat.

Tested with CMake 4.1.2.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>